Return cards with deck

+++ Super Cartes Infinies/Services/DeckService.cs
@@ -62,7 +62,12 @@ namespace Super_Cartes_Infinies.Services
         public async Task<ActionResult<String>> DeleteDeck(int deckId, string userId)
         {
             Player currentPlayer = await _context.Players.Where(x => x.IdentityUserId == userId).FirstOrDefaultAsync();
-            Deck deleteDeck = await _context.Decks.Where(x => x.PlayerId == currentPlayer.Id).Where(x => x.Id == deckId).FirstOrDefaultAsync();
+            Deck deleteDeck = await _context.Decks.Where(x => x.Id == deckId).FirstOrDefaultAsync();
+
+            if(deleteDeck.PlayerId != currentPlayer.Id)
+            {
+                return "This is not your deck.";
+            }
 
             _context.Decks.Remove(deleteDeck);
             _context.SaveChangesAsync();
